requirements.txt
Django==4.1.5
gunicorn==20.1.0
whitenoise==6.3.0

psycopg2==2.9.5

redis==4.4.2
celery==5.2.7

flake8==6.0.0
isort==5.11.4
black==22.12.0
djangorestframework